\name{block.SRM.generalized.reciprocity}

\alias{block.SRM.generalized.reciprocity}

\title{Computing generalized reciprocity in SRM block designs}

\description{
A function for computing generalized reciprocity in SRM block designs.
}


\usage{
   block.SRM.generalized.reciprocity(X)
}

\arguments{
  \item{X}{Original block sociomatrix.}}

\details{
\code{block.SRM.generalized.reciprocity} estimates generalized reciprocity for SRM block designs.
}

\value{
\item{generalized.reciprocity}{Generalized reciprocity measure for SRM block designs.}
}

\references{

Kenny, D. A., Kashy, D. A., & Cook, W. L. (2006). \emph{Dyadic data analysis}. New York: Guilford Press.
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu> & Lluís Salafranca <lsalafranca@ub.edu>.
}

\seealso{ \code{\link{block.SRM.variances}}
}

\keyword{htest}
