\name{nonindependence-package}
\alias{nonindependence-package}
\alias{nonindependence}
\docType{package}
\title{
Quantifying and testing nonindependence in standard dyadic and SRM designs
}
\description{
nonindependence is a package that computes interdependence in several types of dyadic data designs.
}
\details{
\tabular{ll}{
Package: \tab nonindependence\cr
Version: \tab 0.1\cr
Date: \tab 2009-06-12\cr
Depends: \tab >= 2.8.1\cr
License: \tab GPL version 2 or newer\cr
}

Index:
\preformatted{
block.SRM.between.groups.tTest				Between groups t test for SRM block designs
block.SRM.dyadic.reciprocity				Dyadic reciprocity in SRM block designs
block.SRM.dyadic.reciprocity.tTest			Dyadic covariance t test in SRM block designs
block.SRM.effects					SRM effects in block designs
block.SRM.generalized.reciprocity			Generalized reciprocity in SRM block designs
block.SRM.jackknife					Jackknife test in SRM block designs
block.SRM.relative.variances				SRM relative variances in block designs
block.SRM.reliability					Reliability of SRM estimators in block designs
block.SRM.variances					SRM absolute variances in block designs
categorical.distinguishable.dyad			Nonindependence for distinguishable dyads and nominal outcomes
categorical.indistinguishable.dyad			Nonindependence for indistinguishable dyads and nominal outcomes
distinguishable.dyad					Nonindependence for distinguishable dyads and interval outcomes
indistinguishable.dyad					Nonindependence for distinguishable dyads and interval outcomes
pairwise.correlation					Pairwise correlation for measuring nonindependence
prepare.block.matrix					Data preparation for analyzing SRM block designs
prepare.SRM.matrix					Data preparation for analyzing SRM round robin designs
SRM.between.groups.tTest				Between groups t test for round robin designs
SRM.dyadic.covariance.Ftest				Dyadic covariance F test in SRM round robin designs
SRM.dyadic.reciprocity					Dyadic reciprocity in SRM round robin designs
SRM.effects						SRM effects in round robin designs
SRM.generalized.reciprocity				Generalized reciprocity in SRM round robin designs
SRM.jackknife						Jackknife test in SRM round robin designs
SRM.relative.variances					SRM relative variances in round robin designs
SRM.reliability						Reliability of SRM estimators in round robin designs
SRM.variances						SRM absolute variances in round robin designs
SRM.within.groups.tTest					Within groups t test for SRM round robin designs
}
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu> & Lluís Salafranca <lsalafranca@ub.edu>.

Maintainer: David Leiva <dleivaur@ub.edu>
}
\references{
Kenny, D. A., Kashy, D. A., & Cook, W. L. (2006). \emph{Dyadic data analysis}. New York: Guilford Press.

Lashley, B. R., & Bond, C. F., Jr. (1997). Significance testing for round robin data. \emph{Psychological Methods}, \emph{2}, 278-291.
}
\keyword{package}
\seealso{
For more information see \code{\link[RcmdrPlugin.nonindependence]{RCmdrPlugin.nonindependence-package}}.
}
