\name{SRM.dyadic.reciprocity}

\alias{SRM.dyadic.reciprocity}

\title{Computing dyadic reciprocity in SRM round robin designs}

\description{
A function for computing dyadic reciprocity in SRM round robin designs.
}


\usage{
   SRM.dyadic.reciprocity(X)
}

\arguments{
  \item{X}{Original round robin sociomatrix.}
  }

\details{
\code{SRM.dyadic.reciprocity} estimates dyadic reciprocity for round robin designs.
}

\value{
\item{dyadic.reciprocity}{Dyadic reciprocity measure for SRM round robin designs.}
}

\references{

Kenny, D. A., Kashy, D. A., & Cook, W. L. (2006). \emph{Dyadic data analysis}. New York: Guilford Press.
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu> & Lluís Salafranca <lsalafranca@ub.edu>.
}

\seealso{ \code{\link{SRM.effects}},\code{\link{SRM.variances}}
}

\keyword{htest}
