\name{SRM.generalized.reciprocity}

\alias{SRM.generalized.reciprocity}

\title{Computing generalized reciprocity in SRM round robin designs}

\description{
A function for computing generalized reciprocity in SRM round robin designs.
}

\usage{
   SRM.generalized.reciprocity(X)
}

\arguments{
  \item{X}{Original round robin sociomatrix.}
}

\details{
\code{SRM.generalized.reciprocity} estimates generalized reciprocity for SRM round robin designs.
}

\value{
\item{generalized.reciprocity}{Generalized reciprocity measure for SRM round robin designs.}
}

\references{

Kenny, D. A., Kashy, D. A., & Cook, W. L. (2006). \emph{Dyadic data analysis}. New York: Guilford Press.
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu> & Lluís Salafranca <lsalafranca@ub.edu>.
}

\seealso{ \code{\link{SRM.variances}}
}

\keyword{htest}
